feat(settings): add credentials section and rewrite descriptions
- Add CREDENTIALS to SECTION_IDS and "credential" to field kind union
- Extract buildDefaultSections to settings-section-defs.ts for modularity
- Add credentials section with inline key management (pills, add/delete)
- Rewrite all 9 section descriptions to remove jargon (audit fix A11)
- Add groupId to all sections for rail group assignment
- Remove unused getRuntimeMetadataFields export (knip fix)
- Add CSS for credential pills, delete buttons, trust message
- Add unit tests for credentials section and jargon blocklist
