+++ CLAUDE.local.md
@@ -1,0 +1,17 @@
+# CLAUDE.local.md
+
+Personal preferences for symphony-orchestrator.
+
+## Role
+
+Project owner/maintainer. Makes architectural decisions and sets coding standards.
+
+## Communication Style
+
+Detailed by default — include context and reasoning alongside answers. Match depth to the task when helpful.
+
+## Notes
+
+- Expert-level technical depth — skip foundational explanations
+- Prefers structured explanations with bullet points and step-by-step instructions
+- Conversational tone over strictly professional
