+++ frontend/src/styles/settings.css
@@ -1,6 +1,6 @@
 .settings-layout {
   display: grid;
-  grid-template-columns: clamp(180px, 20vw, 240px) minmax(0, 1fr);
+  grid-template-columns: 240px minmax(0, 1fr);
   align-items: start;
   gap: var(--space-8);
 }
@@ -54,7 +54,6 @@
   align-items: flex-start;
   gap: 2px;
   width: 100%;
-  min-height: 44px;
   padding: var(--space-2) var(--space-3);
   border: none;
   border-left: 2px solid transparent;
@@ -70,13 +69,13 @@
 }
 
 .settings-nav-item:hover {
-  background: var(--bg-subtle);
+  background: var(--bg-subtle, #f1f5f9);
   color: var(--text-primary);
 }
 
 .settings-nav-item.is-selected {
-  border-left-color: var(--interactive-primary);
-  background: color-mix(in srgb, var(--interactive-primary) 6%, var(--bg-surface));
+  border-left-color: var(--interactive-primary, #0f766e);
+  background: color-mix(in srgb, var(--interactive-primary, #0f766e) 6%, var(--bg-surface));
   color: var(--text-primary);
 }
 
@@ -94,7 +93,7 @@
 }
 
 .settings-nav-item.is-selected .settings-nav-title {
-  color: var(--interactive-primary);
+  color: var(--interactive-primary, #0f766e);
   font-weight: 600;
 }
 
@@ -110,44 +109,8 @@
 
 .settings-nav-separator {
   height: 1px;
-  background: var(--border-subtle);
+  background: var(--border-subtle, #e2e8f0);
   margin: var(--space-2);
-}
-
-/* ── Group headers in rail ─────────────────────────────── */
-.settings-nav-group-header {
-  display: flex;
-  align-items: center;
-  gap: var(--space-2);
-  padding: var(--space-3) var(--space-3) var(--space-1);
-  color: var(--text-muted);
-  font-size: var(--text-xs);
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: var(--tracking-wider);
-  margin-top: var(--space-3);
-}
-
-.settings-nav-group-header:first-child {
-  margin-top: 0;
-}
-
-.settings-nav-group-header svg {
-  width: 14px;
-  height: 14px;
-  flex-shrink: 0;
-}
-
-/* ── Start Here badge ──────────────────────────────────── */
-.settings-start-here {
-  display: inline-flex;
-  align-items: center;
-  padding: 1px var(--space-2);
-  border-radius: var(--radius-full);
-  background: color-mix(in srgb, var(--interactive-primary) 15%, transparent);
-  color: var(--interactive-primary);
-  font-size: var(--text-xs);
-  font-weight: 600;
 }
 
 .settings-content {
@@ -205,7 +168,6 @@
   padding: var(--space-5) var(--space-6);
   background: var(--bg-surface);
   animation: settings-card-in 150ms ease both;
-  will-change: opacity, transform;
 }
 
 @keyframes settings-card-in {
@@ -266,7 +228,7 @@
 }
 
 .settings-group-collapsed {
-  border-top: 1px solid var(--border-subtle);
+  border-top: 1px solid var(--border-subtle, #e2e8f0);
   padding-top: var(--space-3);
 }
 
@@ -304,7 +266,7 @@
 
 .settings-grid {
   display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(340px, 1fr));
+  grid-template-columns: repeat(2, minmax(0, 1fr));
   gap: var(--space-5) var(--space-5);
   align-items: start;
 }
@@ -362,7 +324,8 @@
 .settings-grid [data-field-kind="json"],
 .settings-grid [data-field-kind="list"],
 .settings-grid [data-field-kind="textarea"],
-.settings-grid [data-field-kind="readonly"] {
+.settings-grid [data-field-kind="readonly"],
+.settings-grid [data-field-kind="credential"] {
   grid-column: 1 / -1;
 }
 
@@ -385,7 +348,7 @@
 .settings-dev-tools {
   margin-top: var(--space-4);
   padding-top: var(--space-3);
-  border-top: 1px dashed var(--border-subtle);
+  border-top: 1px dashed var(--border-subtle, #e2e8f0);
 }
 
 .settings-dev-tools > summary {
@@ -431,7 +394,7 @@
   flex-wrap: wrap;
   gap: var(--space-2);
   padding: var(--space-3);
-  background: var(--bg-subtle);
+  background: var(--bg-subtle, #f1f5f9);
   border-radius: var(--radius-md, 6px);
 }
 
@@ -441,7 +404,7 @@
   max-height: 320px;
   overflow-y: auto;
   font-size: var(--text-xs);
-  background: var(--bg-subtle);
+  background: var(--bg-subtle, #f1f5f9);
   border-radius: var(--radius-md, 6px);
   padding: var(--space-3);
 }
@@ -479,7 +442,7 @@
   max-height: 420px;
   display: flex;
   flex-direction: column;
-  border: 1px solid var(--border-subtle);
+  border: 1px solid var(--border-subtle, #e2e8f0);
   border-radius: var(--radius-lg, 10px);
   background: var(--bg-surface);
   box-shadow: var(--shadow-xl);
@@ -491,7 +454,7 @@
   font-size: var(--text-sm);
   font-weight: 600;
   color: var(--text-primary);
-  border-bottom: 1px solid var(--border-subtle);
+  border-bottom: 1px solid var(--border-subtle, #e2e8f0);
 }
 
 .project-picker-list {
@@ -521,7 +484,7 @@
 
 .project-picker-card:hover {
   background: var(--bg-elevated);
-  border-color: var(--interactive-primary);
+  border-color: var(--interactive-primary, #0f766e);
 }
 
 .project-picker-slug {
@@ -530,13 +493,52 @@
   font-family: var(--font-mono, monospace);
 }
 
-.settings-nav-item:focus-visible {
+/* ── Credential field ──────────────────────────────────── */
+.settings-credential-pills {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--space-2);
+}
+
+.settings-credential-pill {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--space-1);
+  padding: var(--space-1) var(--space-2);
+  background: var(--bg-subtle);
+  border: 1px solid var(--border-subtle);
+  border-radius: var(--radius-full);
+  font-size: var(--text-sm);
+  font-family: var(--font-mono);
+}
+
+.settings-credential-delete {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: 18px;
+  height: 18px;
+  padding: 0;
+  border: none;
+  background: none;
+  color: var(--text-muted);
+  cursor: pointer;
+  border-radius: 50%;
+  font-size: var(--text-xs);
+}
+
+.settings-credential-delete:hover {
+  background: var(--bg-subtle);
+  color: var(--status-blocked);
+}
+
+.settings-credential-delete:focus-visible {
   outline: 2px solid var(--interactive-primary);
   outline-offset: 2px;
 }
 
-.settings-actions button:focus-visible,
-.settings-dev-tools summary:focus-visible {
-  outline: 2px solid var(--interactive-primary);
-  outline-offset: 2px;
+.settings-credential-trust {
+  margin-top: var(--space-3);
+  font-size: var(--text-xs);
+  font-style: italic;
 }
