+++ frontend/src/views/settings-helpers.ts
@@ -1,6 +1,6 @@
 import type { IconName } from "../ui/icons";
-import { REASONING_EFFORT_OPTIONS } from "../types";
 import { buildDiffText, prettyJson, redactPath } from "./config-helpers";
+import { buildDefaultSections } from "./settings-section-defs.js";
 import { buildSectionPatchPlan } from "./settings-patches";
 import { getValueAtPath, setValueAtPath } from "./settings-paths";
 
@@ -15,6 +15,7 @@ export const SECTION_IDS = {
   WORKFLOW_STAGES: "workflow-stages",
   FEATURE_FLAGS: "feature-flags",
   RUNTIME_PATHS: "runtime-paths",
+  CREDENTIALS: "credentials",
 } as const;
 
 export const SECTION_GROUPS = {
@@ -42,7 +43,7 @@ export interface SettingsFieldOption {
 export interface SettingsFieldDefinition {
   path: string;
   label: string;
-  kind: "text" | "number" | "textarea" | "select" | "boolean" | "json" | "list" | "readonly";
+  kind: "text" | "number" | "textarea" | "select" | "boolean" | "json" | "list" | "readonly" | "credential";
   group?: string;
   groupDescription?: string;
   advanced?: boolean;
@@ -197,24 +198,6 @@ export function buildUnderlyingPaths(section: SettingsSectionDefinition): string
   return section.fields.map((field) => field.path);
 }
 
-export function getRuntimeMetadataFields(effective: Record<string, unknown>): SettingsFieldDefinition[] {
-  const metadataPaths = [
-    ["runtime.workflow_path", "Workflow path"],
-    ["runtime.data_dir", "Data directory"],
-    ["runtime.provider_summary", "Provider summary"],
-  ] as const;
-  return metadataPaths.map(([path, label]) => ({
-    path,
-    label,
-    kind: "readonly",
-    hint:
-      getValueAtPath(effective, path) === undefined
-        ? "Runtime metadata not exposed yet."
-        : "Read-only runtime metadata for operator confidence.",
-    editable: false,
-  }));
-}
-
 export function getSectionById(
   schema: Record<string, unknown> | null,
   effective: Record<string, unknown>,
@@ -310,386 +293,6 @@ function buildSchemaSections(schema: Record<string, unknown> | null): SettingsSe
     });
   });
   return parsedSections;
-}
-
-function buildDefaultSections(effective: Record<string, unknown>): SettingsSectionDefinition[] {
-  const featureFlagPath = getValueAtPath(effective, "feature_flags") !== undefined ? "feature_flags" : "flags";
-  return [
-    {
-      id: "tracker",
-      title: "Tracker",
-      description: "Connect Symphony to your issue tracker and define which states mean 'in progress' vs 'done'.",
-      badge: "Start here",
-      prefixes: ["tracker"],
-      saveLabel: "Save tracker",
-      groupId: SECTION_GROUPS.SETUP.id,
-      startHere: true,
-      fields: [
-        {
-          path: "tracker.kind",
-          label: "Issue tracker",
-          kind: "select",
-          group: "Connection",
-          groupDescription: "Choose the tracker Symphony should read work from.",
-          options: [{ value: "linear", label: "Linear" }],
-        },
-        {
-          path: "tracker.endpoint",
-          label: "Endpoint override",
-          kind: "text",
-          group: "Advanced connection",
-          groupDescription: "Only change this if you route tracker traffic through a custom endpoint.",
-          advanced: true,
-          hint: "Leave blank to use Linear's default GraphQL endpoint.",
-        },
-        {
-          path: "tracker.project_slug",
-          label: "Linear project",
-          kind: "text",
-          group: "Connection",
-          hint: "Pick the Linear project Symphony should dispatch from.",
-          actionLabel: "Browse",
-          actionKind: "browse-linear-projects",
-        },
-        {
-          path: "tracker.active_states",
-          label: "States that mean work is active",
-          kind: "list",
-          group: "Workflow meaning",
-          groupDescription: "Tell Symphony which tracker states mean an issue is ready or in progress.",
-          hint: "One state per line. Example: Todo, In Progress.",
-          placeholder: "Todo\nIn Progress",
-        },
-        {
-          path: "tracker.terminal_states",
-          label: "States that mean work is done",
-          kind: "list",
-          group: "Workflow meaning",
-          hint: "One state per line. Example: Done, Canceled.",
-          placeholder: "Done\nCanceled",
-        },
-      ],
-    },
-    {
-      id: "model-provider-auth",
-      title: "Model provider / auth",
-      description:
-        "Choose the default model, decide how Symphony authenticates, and override provider routing if needed.",
-      badge: "Core runtime",
-      prefixes: ["codex.model", "codex.reasoning_effort", "codex.auth", "codex.provider"],
-      saveLabel: "Save provider settings",
-      groupId: SECTION_GROUPS.AGENT.id,
-      fields: [
-        {
-          path: "codex.model",
-          label: "Default model",
-          kind: "text",
-          group: "Default model",
-          groupDescription: "Set the model Symphony should use unless an issue override is saved.",
-        },
-        {
-          path: "codex.reasoning_effort",
-          label: "Reasoning effort",
-          kind: "select",
-          group: "Default model",
-          options: REASONING_EFFORT_OPTIONS.map((value) => ({ value, label: value })),
-        },
-        {
-          path: "codex.auth.mode",
-          label: "How to authenticate",
-          kind: "select",
-          group: "Authentication",
-          groupDescription: "Use an API key or your local OpenAI login, depending on how you run Codex.",
-          options: [
-            { value: "api_key", label: "API key" },
-            { value: "openai_login", label: "OpenAI login" },
-          ],
-        },
-        {
-          path: "codex.auth.source_home",
-          label: "Auth directory",
-          kind: "text",
-          group: "Authentication",
-          advanced: true,
-          hint: "Local directory where auth credentials are stored.",
-        },
-        {
-          path: "codex.provider.id",
-          label: "Provider id",
-          kind: "text",
-          group: "Provider routing",
-          groupDescription: "Only fill these in when routing Codex through a custom or OpenAI-compatible provider.",
-          advanced: true,
-          placeholder: "custom",
-        },
-        {
-          path: "codex.provider.name",
-          label: "Provider name",
-          kind: "text",
-          group: "Provider routing",
-          advanced: true,
-        },
-        {
-          path: "codex.provider.base_url",
-          label: "Base URL",
-          kind: "text",
-          group: "Provider routing",
-          advanced: true,
-        },
-        {
-          path: "codex.provider.env_key",
-          label: "API key env var",
-          kind: "text",
-          group: "Provider routing",
-          advanced: true,
-          redact: true,
-          hint: "Name of the environment variable holding your API key.",
-        },
-        {
-          path: "codex.provider.wire_api",
-          label: "API protocol",
-          kind: "text",
-          group: "Provider routing",
-          advanced: true,
-          placeholder: "responses",
-          hint: "Wire protocol used to communicate with this provider (e.g. responses).",
-        },
-        {
-          path: "codex.provider.requires_openai_auth",
-          label: "Requires OpenAI authentication",
-          kind: "boolean",
-          group: "Provider routing",
-          advanced: true,
-        },
-        {
-          path: "codex.provider.env_http_headers",
-          label: "Custom HTTP header env vars",
-          kind: "json",
-          group: "Provider routing",
-          advanced: true,
-          hint: "Map of HTTP header names to environment variable names.",
-        },
-      ],
-    },
-    {
-      id: "sandbox",
-      title: "Sandbox",
-      description: "Container image, safety posture, and resource controls for each Codex turn.",
-      badge: "Execution safety",
-      prefixes: ["codex.approval_policy", "codex.thread_sandbox", "codex.turn_sandbox_policy", "codex.sandbox"],
-      saveLabel: "Save sandbox",
-      groupId: SECTION_GROUPS.AGENT.id,
-      fields: [
-        {
-          path: "codex.approval_policy",
-          label: "Approval policy",
-          kind: "json",
-          group: "Safety",
-          groupDescription: "Decide which actions need explicit approval before they run.",
-          hint: "Controls which operations require human approval before execution.",
-        },
-        {
-          path: "codex.thread_sandbox",
-          label: "Thread sandbox",
-          kind: "text",
-          group: "Container",
-          groupDescription: "Choose the default sandbox mode and container image for each turn.",
-        },
-        {
-          path: "codex.turn_sandbox_policy",
-          label: "Turn sandbox policy",
-          kind: "json",
-          group: "Safety",
-          advanced: true,
-          hint: "Per-turn policy overriding the default sandbox behavior.",
-        },
-        { path: "codex.sandbox.image", label: "Sandbox image", kind: "text", group: "Container" },
-        {
-          path: "codex.sandbox.network",
-          label: "Network access",
-          kind: "text",
-          group: "Container",
-          advanced: true,
-        },
-        {
-          path: "codex.sandbox.security.no_new_privileges",
-          label: "No new privileges",
-          kind: "boolean",
-          group: "Safety",
-          advanced: true,
-        },
-        {
-          path: "codex.sandbox.security.drop_capabilities",
-          label: "Drop capabilities",
-          kind: "boolean",
-          group: "Safety",
-          advanced: true,
-        },
-        {
-          path: "codex.sandbox.security.gvisor",
-          label: "Use gVisor",
-          kind: "boolean",
-          group: "Safety",
-          advanced: true,
-        },
-        {
-          path: "codex.sandbox.resources.memory",
-          label: "Memory",
-          kind: "text",
-          group: "Resources",
-          groupDescription: "Optional resource caps for each sandboxed run.",
-          advanced: true,
-        },
-        { path: "codex.sandbox.resources.cpus", label: "CPUs", kind: "text", group: "Resources", advanced: true },
-        {
-          path: "codex.sandbox.resources.tmpfs_size",
-          label: "Tmpfs size",
-          kind: "text",
-          group: "Resources",
-          advanced: true,
-        },
-      ],
-    },
-    {
-      id: "agent",
-      title: "Agent",
-      description: "Concurrency, retry limits, and continuation controls for worker agents.",
-      badge: "Core runtime",
-      prefixes: ["agent"],
-      saveLabel: "Save agent settings",
-      groupId: SECTION_GROUPS.AGENT.id,
-      fields: [
-        {
-          path: "agent.max_concurrent_agents",
-          label: "Max concurrent agents",
-          kind: "number",
-          group: "Throughput",
-          groupDescription: "Control how much work Symphony runs in parallel.",
-        },
-        { path: "agent.max_turns", label: "Max turns", kind: "number", group: "Throughput" },
-        {
-          path: "agent.max_continuation_attempts",
-          label: "Max continuation attempts",
-          kind: "number",
-          group: "Continuation",
-          groupDescription: "Prevent long-running issues from looping forever without a stop signal.",
-          hint: "How many times an agent can retry without emitting a stop signal before being stopped. Default: 5.",
-        },
-        {
-          path: "agent.max_retry_backoff_ms",
-          label: "Max retry backoff (ms)",
-          kind: "number",
-          group: "Retries",
-          groupDescription: "Tune how long Symphony waits before retrying failed work.",
-          advanced: true,
-        },
-      ],
-    },
-    {
-      id: "repositories-github",
-      title: "Repositories / GitHub",
-      description: "Shaped views over repo routing and GitHub automation config — no extra hidden state.",
-      badge: "Integration",
-      prefixes: ["repos", "github"],
-      saveLabel: "Save repository settings",
-      groupId: SECTION_GROUPS.SETUP.id,
-      fields: [
-        { path: "repos", label: "Repository routing", kind: "json", hint: "JSON array of repo routing rules." },
-        { path: "github.api_base_url", label: "GitHub API base URL", kind: "text" },
-        {
-          path: "github.token",
-          label: "GitHub token",
-          kind: "text",
-          redact: true,
-          hint: "Redacted after load for safety.",
-        },
-      ],
-    },
-    {
-      id: "notifications",
-      title: "Notifications",
-      description: "Slack delivery settings surfaced directly from notifications config.",
-      badge: "Integration",
-      prefixes: ["notifications"],
-      saveLabel: "Save notifications",
-      groupId: SECTION_GROUPS.NOTIFICATIONS.id,
-      fields: [
-        {
-          path: "notifications.slack.verbosity",
-          label: "Slack verbosity",
-          kind: "select",
-          options: ["off", "critical", "verbose"].map((value) => ({ value, label: value })),
-        },
-        {
-          path: "notifications.slack.webhook_url",
-          label: "Slack webhook URL",
-          kind: "text",
-          redact: true,
-          hint: "Existing webhook values stay redacted. Enter a new one to replace it.",
-        },
-      ],
-    },
-    {
-      id: "workflow-stages",
-      title: "Workflow stages",
-      description: "Stage machine structure for operators who want explicit pipeline control.",
-      badge: "Advanced",
-      prefixes: ["state_machine"],
-      saveLabel: "Save workflow stages",
-      groupId: SECTION_GROUPS.SYSTEM.id,
-      fields: [
-        { path: "state_machine.stages", label: "Stages", kind: "json", hint: "JSON array of stage definitions." },
-        {
-          path: "state_machine.transitions",
-          label: "Transitions",
-          kind: "json",
-          hint: "JSON object keyed by source stage.",
-        },
-      ],
-    },
-    {
-      id: "feature-flags",
-      title: "Feature flags",
-      description: "Operator-facing flag overlay if the runtime exposes flags in config. Otherwise we say so plainly.",
-      badge: "Runtime dependent",
-      prefixes: [featureFlagPath],
-      saveLabel: "Save flags",
-      groupId: SECTION_GROUPS.SYSTEM.id,
-      fields: [
-        getValueAtPath(effective, featureFlagPath) === undefined
-          ? {
-              path: "runtime.feature_flags",
-              label: "Feature flags",
-              kind: "readonly",
-              hint: "Feature flag metadata not exposed yet.",
-              editable: false,
-            }
-          : {
-              path: featureFlagPath,
-              label: "Feature flags",
-              kind: "json",
-              hint: "JSON object keyed by feature flag name.",
-            },
-      ],
-    },
-    {
-      id: "runtime-paths",
-      title: "Runtime / paths",
-      description:
-        "Where Symphony runs, stores workspaces, and listens locally. Missing metadata is shown as not exposed yet.",
-      badge: "Runtime info",
-      prefixes: ["workspace", "hooks", "polling", "server", "runtime"],
-      saveLabel: "Save runtime settings",
-      groupId: SECTION_GROUPS.SYSTEM.id,
-      fields: [
-        { path: "workspace.root", label: "Workspace root", kind: "text" },
-        { path: "hooks.timeout_ms", label: "Hook timeout (ms)", kind: "number" },
-        { path: "polling.interval_ms", label: "Polling interval (ms)", kind: "number" },
-        { path: "server.port", label: "Server port", kind: "number" },
-        ...getRuntimeMetadataFields(effective),
-      ],
-    },
-  ];
 }
 
 function pickByPrefixes(source: Record<string, unknown>, prefixes: string[]): Record<string, unknown> {
