+++ skills/visual-verify/visual-verify
@@ -1,0 +1,1 @@
+../../skills/visual-verify
